replaced magic value in play result

+++ RockPaperScissors.UnitTest/RoundTest.cs
@@ -22,30 +22,30 @@ namespace RockPaperScissors.UnitTest
         [Fact]
         public void PaperShouldWrapRock()
         {
-            target.Play("Paper", "Rock").Should().Be(1);
-            target.Play("Rock", "Paper").Should().Be(2);
+            ((int) target.Play("Paper", "Rock")).Should().Be(1);
+            ((int) target.Play("Rock", "Paper")).Should().Be(2);
         }
 
         [Fact]
         public void RockShouldBluntScissors()
         {
-            target.Play("Rock", "Scissors").Should().Be(1);
-            target.Play("Scissors", "Rock").Should().Be(2);
+            ((int) target.Play("Rock", "Scissors")).Should().Be(1);
+            ((int) target.Play("Scissors", "Rock")).Should().Be(2);
         }
 
         [Fact]
         public void SameChoiceShouldResultInDraw()
         {
-            target.Play("Rock", "Rock").Should().Be(0);
-            target.Play("Paper", "Paper").Should().Be(0);
-            target.Play("Scissors", "Scissors").Should().Be(0);
+            ((int) target.Play("Rock", "Rock")).Should().Be(0);
+            ((int) target.Play("Paper", "Paper")).Should().Be(0);
+            ((int) target.Play("Scissors", "Scissors")).Should().Be(0);
         }
 
         [Fact]
         public void ScissorsShouldCutPaper()
         {
-            target.Play("Scissors", "Paper").Should().Be(1);
-            target.Play("Paper", "Scissors").Should().Be(2);
+            ((int) target.Play("Scissors", "Paper")).Should().Be(1);
+            ((int) target.Play("Paper", "Scissors")).Should().Be(2);
         }
     }
 }
